relocate stop functionality
