+++ embark/dashboard/forms.py
@@ -1,0 +1,7 @@
+import logging
+from django import forms
+
+logger = logging.getLogger('web')
+
+class StopAnalysisForm(forms.Form):
+    id = forms.UUIDField
