+++ embark/dashboard/urls.py
@@ -9,4 +9,5 @@ urlpatterns = [
     path('dashboard/service/', views.service_dashboard, name='embark-ServiceDashboard'),
     path('dashboard/report/', views.report_dashboard, name='embark-ReportDashboard'),
     path('dashboard/individualReport/<uuid:analysis_id>', views.individual_report_dashboard, name='embark-IndividualReportDashboard'),
+    path('dashboard/stop/', views.stop_analysis, name='embark-stop-analysis')
 ]
