+++ embark/templates/dashboard/serviceDashboard.html
@@ -5,5 +5,26 @@
 {% block title %}EMBArk service dashboard{% endblock title %}
 {% block navigation %}{% include "navigation.html" %}{% endblock navigation %}
 {% block maincontent %}
-    {% csrf_token %}
+{% if success_message %}
+    <div class="alert alert-success alert-dismissible fade show" role="alert">
+        <medium>{{ message }}</medium>
+    </div>
+{% endif %}
+    <div class="row StopRow">
+            <div class="box">
+                <p class="mainText">Select and stop Analysis</p>
+                <form action="{% url 'embark-stop-analysis' %}" method="post" id="stop-form">
+                    <!--  onsubmit="window.confirmDelete(event)" -->
+                    {% csrf_token %}
+                    <div class="innerBlock">
+                        <div id="analysisToStop" class="form-group stopAnalysis" >
+                            {{ form }}
+                        </div>
+                    </div>
+                    <div class="buttonRow">
+                        <button type="stop" class="btn buttonRowElem">Stop</button>
+                    </div>
+                </form>
+            </div>
+    </div>
 {% endblock maincontent %}
